Supported working directories and startup commands when starting.
In the profile, per window, the 'pane' variable is now a list of dicts
instead of a number. each dict may contain a 'cmd' and a 'cwd'
parameter - the shell in that pane will then change directory to 'cwd',
and execute 'cmd'.

This will break all existing profiles :)

Support for saving working directories and running processes is in
the works.